Included the New Class Ball
The new class ball has the ability to create another point in space that act as a repelent. The position of this point is determined by linear equation. thus the rect between the two points always goes to the goal.
And as the ball is an atractor, the attacker will always push the ball in the direction of the goal