Strategist VectorialFieldFunctions Created
All the EstrategistCpp has been actualized to run in cmake, thus making easier for anyone to run the code.
In the other side, the recunstruciton of the proyect also led to changes in the structure of the code. Thus making it more scalable (the modeling for the vectorial field entities were based in some way by unity )